mrc-capturer-absorb(shared-mooncake-gate-move,mechanical_provable): Move maybe_init_shared_mooncake_transfer_engine to mooncake_transfer_engine (cut+paste)
