+++ python/sglang/srt/model_executor/model_runner.py
@@ -38,7 +38,7 @@ from sglang.srt.distributed import (
     get_world_group,
 )
 from sglang.srt.distributed.device_communicators.mooncake_transfer_engine import (
-    init_mooncake_transfer_engine,
+    maybe_init_shared_mooncake_transfer_engine,
 )
 from sglang.srt.distributed.device_communicators.pynccl_allocator import (
     prealloc_symmetric_memory_pool,
@@ -197,7 +197,6 @@ from sglang.srt.utils import (
     set_cuda_arch,
     slow_rank_detector,
 )
-from sglang.srt.utils.network import get_local_ip_auto
 from sglang.srt.utils.nvtx_pytorch_hooks import PytHooks
 from sglang.srt.utils.nvtx_utils import profile_range
 from sglang.srt.utils.offloader import (
@@ -1830,46 +1829,3 @@ class ModelRunner(ModelRunnerKVCacheMixin):
             load_format=load_format,
         )
         self.load_config = load_config
-
-
-def maybe_init_shared_mooncake_transfer_engine(
-    *, server_args: ServerArgs, gpu_id: int
-) -> None:
-    """
-    Need MooncakeTransferEngine when:
-    1) PD disaggregation uses mooncake for KV transfer (prefill/decode)
-    2) HiCache uses mooncake storage backend
-    3) Encoder disaggregation uses mooncake
-    """
-    use_mooncake_te = (
-        (
-            server_args.disaggregation_mode != "null"
-            and server_args.disaggregation_transfer_backend == "mooncake"
-        )
-        or (
-            server_args.enable_hierarchical_cache
-            and server_args.hicache_storage_backend == "mooncake"
-            and envs.SGLANG_HICACHE_MOONCAKE_REUSE_TE.get()
-        )
-        or (
-            server_args.encoder_only
-            and server_args.encoder_transfer_backend == "mooncake"
-        )
-        or (
-            server_args.language_only
-            and server_args.encoder_transfer_backend == "mooncake"
-        )
-        or (
-            server_args.enable_elastic_expert_backup
-            and server_args.elastic_ep_backend is not None
-        )
-    )
-
-    if use_mooncake_te:
-        init_mooncake_transfer_engine(
-            hostname=get_local_ip_auto(),
-            gpu_id=gpu_id,
-            ib_device=(
-                server_args.disaggregation_ib_device or server_args.mooncake_ib_device
-            ),
-        )
